chore: relocate dist/components to dist/atomic/components
https://coveord.atlassian.net/browse/KIT-3858
