+++ packages/atomic/stencil-proxy/atomic/atomic.esm.js
@@ -1,4 +1,4 @@
 import './autoloader/index.esm.js';
 
 export * from './_atomic.esm.js';
-export * from '../components/components/lazy-index.js';
+export * from './components/components/lazy-index.js';
