+++ packages/atomic/stencil-proxy/esm/index.js
@@ -1,4 +1,4 @@
 import '../atomic/autoloader/index.esm.js';
 
 export * from './_index.js';
-export * as SearchComponentMap from '../components/components/search/lazy-index.js';
+export * as SearchComponentMap from '../atomic/components/components/search/lazy-index.js';
